requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
SQLAlchemy>=2.0.30
psycopg[binary]>=3.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
Authlib>=1.3.0
requests>=2.31.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
bleach>=6.1.0
python-slugify>=8.0.0
Flask-Limiter>=3.5.0
Flask-Caching>=2.1.0
Flask-SocketIO>=5.3.0
eventlet>=0.36.1
redis>=5.0.0
python-socketio>=5.10.0
pygments>=2.17.0
Flask-Babel>=4.0.0